feat: should allow that only the creator can delete HT-3

+++ app/Http/Controllers/Question/DeleteController.php
@@ -4,15 +4,13 @@ namespace App\Http\Controllers\Question;
 
 use App\Http\Controllers\Controller;
 use App\Models\Question;
-use Illuminate\Http\Request;
 
 class DeleteController extends Controller
 {
-    /**
-     * Handle the incoming request.
-     */
     public function __invoke(Question $question)
     {
+        $this->authorize('forceDelete', $question);
+
         $question->forceDelete();
 
         return response()->noContent();
